Add files via upload
This code provides a framework for conducting data-driven classification models using Random Forest (RF) and XGBoost (XGB) algorithms. The case study in this code is applied to an industrial dataset for determination of a pass or failed flushing process. Aside visualizing the performance metrics of each model through a confusion matrix, the code provides insights to analyzing effects of each feature on the output of the model through permutation feature importance.